black

+++ xarray/tests/test_accessor_str.py
@@ -834,15 +834,11 @@ def test_extract_multi_nocase(dtype):
 
 
 def test_extract_broadcast(dtype):
-    value = xr.DataArray(
-        ["a_Xy_0", "ab_xY_10", "abc_Xy_01"],
-        dims=["X"],
-    ).astype(dtype)
+    value = xr.DataArray(["a_Xy_0", "ab_xY_10", "abc_Xy_01"], dims=["X"],).astype(dtype)
 
-    pat_str = xr.DataArray(
-        [r"(\w+)_Xy_(\d*)", r"(\w+)_xY_(\d*)"],
-        dims=["Y"],
-    ).astype(dtype)
+    pat_str = xr.DataArray([r"(\w+)_Xy_(\d*)", r"(\w+)_xY_(\d*)"], dims=["Y"],).astype(
+        dtype
+    )
     pat_re = value.str._re_compile(pat=pat_str)
 
     expected = [
@@ -1165,15 +1161,11 @@ def test_extractall_multi_multi_nocase(dtype):
 
 
 def test_extractall_broadcast(dtype):
-    value = xr.DataArray(
-        ["a_Xy_0", "ab_xY_10", "abc_Xy_01"],
-        dims=["X"],
-    ).astype(dtype)
+    value = xr.DataArray(["a_Xy_0", "ab_xY_10", "abc_Xy_01"], dims=["X"],).astype(dtype)
 
-    pat_str = xr.DataArray(
-        [r"(\w+)_Xy_(\d*)", r"(\w+)_xY_(\d*)"],
-        dims=["Y"],
-    ).astype(dtype)
+    pat_str = xr.DataArray([r"(\w+)_Xy_(\d*)", r"(\w+)_xY_(\d*)"], dims=["Y"],).astype(
+        dtype
+    )
     pat_re = value.str._re_compile(pat=pat_str)
 
     expected = [
@@ -1262,11 +1254,7 @@ def test_findall_single_multi_case(dtype):
 
     expected = [
         [["a"], ["bab", "baab"], ["abc", "cbc"]],
-        [
-            ["abcd", "dcd", "dccd"],
-            [],
-            ["abcdef", "fef"],
-        ],
+        [["abcd", "dcd", "dccd"], [], ["abcdef", "fef"],],
     ]
     expected = [[[dtype(x) for x in y] for y in z] for z in expected]
     expected = np.array(expected, dtype=np.object_)
@@ -1302,16 +1290,8 @@ def test_findall_single_multi_nocase(dtype):
     ).astype(dtype)
 
     expected = [
-        [
-            ["a"],
-            ["ab", "bab", "baab"],
-            ["abc", "cbc"],
-        ],
-        [
-            ["abcd", "dcd", "dccd"],
-            [],
-            ["abcdef", "fef"],
-        ],
+        [["a"], ["ab", "bab", "baab"], ["abc", "cbc"],],
+        [["abcd", "dcd", "dccd"], [], ["abcdef", "fef"],],
     ]
     expected = [[[dtype(x) for x in y] for y in z] for z in expected]
     expected = np.array(expected, dtype=np.object_)
@@ -1472,15 +1452,11 @@ def test_findall_multi_multi_nocase(dtype):
 
 
 def test_findall_broadcast(dtype):
-    value = xr.DataArray(
-        ["a_Xy_0", "ab_xY_10", "abc_Xy_01"],
-        dims=["X"],
-    ).astype(dtype)
+    value = xr.DataArray(["a_Xy_0", "ab_xY_10", "abc_Xy_01"], dims=["X"],).astype(dtype)
 
-    pat_str = xr.DataArray(
-        [r"(\w+)_Xy_\d*", r"\w+_Xy_(\d*)"],
-        dims=["Y"],
-    ).astype(dtype)
+    pat_str = xr.DataArray([r"(\w+)_Xy_\d*", r"\w+_Xy_(\d*)"], dims=["Y"],).astype(
+        dtype
+    )
     pat_re = value.str._re_compile(pat=pat_str)
 
     expected = [[["a"], ["0"]], [[], []], [["abc"], ["01"]]]
@@ -2429,11 +2405,7 @@ def test_partition_whitespace(dtype):
     ).astype(dtype)
 
     exp_part_dim = [
-        [
-            ["abc", " ", "def"],
-            ["spam", " ", "eggs swallow"],
-            ["red_blue", "", ""],
-        ],
+        [["abc", " ", "def"], ["spam", " ", "eggs swallow"], ["red_blue", "", ""],],
         [
             ["test0", " ", "test1 test2 test3"],
             ["", "", ""],
@@ -2442,11 +2414,7 @@ def test_partition_whitespace(dtype):
     ]
 
     exp_rpart_dim = [
-        [
-            ["abc", " ", "def"],
-            ["spam eggs", " ", "swallow"],
-            ["", "", "red_blue"],
-        ],
+        [["abc", " ", "def"], ["spam eggs", " ", "swallow"], ["", "", "red_blue"],],
         [
             ["test0 test1 test2", " ", "test3"],
             ["", "", ""],
@@ -2477,11 +2445,7 @@ def test_partition_comma(dtype):
     ).astype(dtype)
 
     exp_part_dim = [
-        [
-            ["abc", ", ", "def"],
-            ["spam", ", ", "eggs, swallow"],
-            ["red_blue", "", ""],
-        ],
+        [["abc", ", ", "def"], ["spam", ", ", "eggs, swallow"], ["red_blue", "", ""],],
         [
             ["test0", ", ", "test1, test2, test3"],
             ["", "", ""],
@@ -2490,11 +2454,7 @@ def test_partition_comma(dtype):
     ]
 
     exp_rpart_dim = [
-        [
-            ["abc", ", ", "def"],
-            ["spam, eggs", ", ", "swallow"],
-            ["", "", "red_blue"],
-        ],
+        [["abc", ", ", "def"], ["spam, eggs", ", ", "swallow"], ["", "", "red_blue"],],
         [
             ["test0, test1, test2", ", ", "test3"],
             ["", "", ""],
@@ -2803,14 +2763,10 @@ def test_split_comma(dtype):
 
 def test_splitters_broadcast(dtype):
     values = xr.DataArray(
-        ["ab cd,de fg", "spam, ,eggs swallow", "red_blue"],
-        dims=["X"],
+        ["ab cd,de fg", "spam, ,eggs swallow", "red_blue"], dims=["X"],
     ).astype(dtype)
 
-    sep = xr.DataArray(
-        [" ", ","],
-        dims=["Y"],
-    ).astype(dtype)
+    sep = xr.DataArray([" ", ","], dims=["Y"],).astype(dtype)
 
     expected_left = xr.DataArray(
         [
@@ -2920,15 +2876,9 @@ def test_get_dummies(dtype):
 
 
 def test_get_dummies_broadcast(dtype):
-    values = xr.DataArray(
-        ["x~x|x~x", "x", "x|x~x", "x~x"],
-        dims=["X"],
-    ).astype(dtype)
+    values = xr.DataArray(["x~x|x~x", "x", "x|x~x", "x~x"], dims=["X"],).astype(dtype)
 
-    sep = xr.DataArray(
-        ["|", "~"],
-        dims=["Y"],
-    ).astype(dtype)
+    sep = xr.DataArray(["|", "~"], dims=["Y"],).astype(dtype)
 
     expected = [
         [[False, False, True], [True, True, False]],
@@ -2958,10 +2908,7 @@ def test_get_dummies_empty(dtype):
 
 
 def test_splitters_empty_str(dtype):
-    values = xr.DataArray(
-        [["", "", ""], ["", "", ""]],
-        dims=["X", "Y"],
-    ).astype(dtype)
+    values = xr.DataArray([["", "", ""], ["", "", ""]], dims=["X", "Y"],).astype(dtype)
 
     targ_partition_dim = xr.DataArray(
         [
@@ -2980,18 +2927,13 @@ def test_splitters_empty_str(dtype):
     ]
     targ_partition_none = np.array(targ_partition_none, dtype=np.object_)
     del targ_partition_none[-1, -1][-1]
-    targ_partition_none = xr.DataArray(
-        targ_partition_none,
-        dims=["X", "Y"],
-    )
+    targ_partition_none = xr.DataArray(targ_partition_none, dims=["X", "Y"],)
 
     targ_split_dim = xr.DataArray(
-        [[[""], [""], [""]], [[""], [""], [""]]],
-        dims=["X", "Y", "ZZ"],
+        [[[""], [""], [""]], [[""], [""], [""]]], dims=["X", "Y", "ZZ"],
     ).astype(dtype)
     targ_split_none = xr.DataArray(
-        np.array([[[], [], []], [[], [], [""]]], dtype=np.object_),
-        dims=["X", "Y"],
+        np.array([[[], [], []], [[], [], [""]]], dtype=np.object_), dims=["X", "Y"],
     )
     del targ_split_none.data[-1, -1][-1]
 
@@ -3034,8 +2976,7 @@ def test_splitters_empty_str(dtype):
 
 def test_cat_str(dtype):
     values_1 = xr.DataArray(
-        [["a", "bb", "cccc"], ["ddddd", "eeee", "fff"]],
-        dims=["X", "Y"],
+        [["a", "bb", "cccc"], ["ddddd", "eeee", "fff"]], dims=["X", "Y"],
     ).astype(dtype)
     values_2 = "111"
 
@@ -3080,12 +3021,10 @@ def test_cat_str(dtype):
 
 def test_cat_uniform(dtype):
     values_1 = xr.DataArray(
-        [["a", "bb", "cccc"], ["ddddd", "eeee", "fff"]],
-        dims=["X", "Y"],
+        [["a", "bb", "cccc"], ["ddddd", "eeee", "fff"]], dims=["X", "Y"],
     ).astype(dtype)
     values_2 = xr.DataArray(
-        [["11111", "222", "33"], ["4", "5555", "66"]],
-        dims=["X", "Y"],
+        [["11111", "222", "33"], ["4", "5555", "66"]], dims=["X", "Y"],
     )
 
     targ_blank = xr.DataArray(
@@ -3129,13 +3068,9 @@ def test_cat_uniform(dtype):
 
 def test_cat_broadcast_right(dtype):
     values_1 = xr.DataArray(
-        [["a", "bb", "cccc"], ["ddddd", "eeee", "fff"]],
-        dims=["X", "Y"],
+        [["a", "bb", "cccc"], ["ddddd", "eeee", "fff"]], dims=["X", "Y"],
     ).astype(dtype)
-    values_2 = xr.DataArray(
-        ["11111", "222", "33"],
-        dims=["Y"],
-    )
+    values_2 = xr.DataArray(["11111", "222", "33"], dims=["Y"],)
 
     targ_blank = xr.DataArray(
         [["a11111", "bb222", "cccc33"], ["ddddd11111", "eeee222", "fff33"]],
@@ -3177,13 +3112,9 @@ def test_cat_broadcast_right(dtype):
 
 
 def test_cat_broadcast_left(dtype):
-    values_1 = xr.DataArray(
-        ["a", "bb", "cccc"],
-        dims=["Y"],
-    ).astype(dtype)
+    values_1 = xr.DataArray(["a", "bb", "cccc"], dims=["Y"],).astype(dtype)
     values_2 = xr.DataArray(
-        [["11111", "222", "33"], ["4", "5555", "66"]],
-        dims=["X", "Y"],
+        [["11111", "222", "33"], ["4", "5555", "66"]], dims=["X", "Y"],
     )
 
     targ_blank = (
@@ -3242,14 +3173,8 @@ def test_cat_broadcast_left(dtype):
 
 
 def test_cat_broadcast_both(dtype):
-    values_1 = xr.DataArray(
-        ["a", "bb", "cccc"],
-        dims=["Y"],
-    ).astype(dtype)
-    values_2 = xr.DataArray(
-        ["11111", "4"],
-        dims=["X"],
-    )
+    values_1 = xr.DataArray(["a", "bb", "cccc"], dims=["Y"],).astype(dtype)
+    values_2 = xr.DataArray(["11111", "4"], dims=["X"],)
 
     targ_blank = (
         xr.DataArray(
@@ -3307,15 +3232,9 @@ def test_cat_broadcast_both(dtype):
 
 
 def test_cat_multi():
-    values_1 = xr.DataArray(
-        ["11111", "4"],
-        dims=["X"],
-    )
+    values_1 = xr.DataArray(["11111", "4"], dims=["X"],)
 
-    values_2 = xr.DataArray(
-        ["a", "bb", "cccc"],
-        dims=["Y"],
-    ).astype(np.bytes_)
+    values_2 = xr.DataArray(["a", "bb", "cccc"], dims=["Y"],).astype(np.bytes_)
 
     values_3 = np.array(3.4)
 
@@ -3323,10 +3242,7 @@ def test_cat_multi():
 
     values_5 = np.array("", dtype=np.unicode_)
 
-    sep = xr.DataArray(
-        [" ", ", "],
-        dims=["ZZ"],
-    ).astype(np.unicode_)
+    sep = xr.DataArray([" ", ", "], dims=["ZZ"],).astype(np.unicode_)
 
     expected = xr.DataArray(
         [
@@ -3366,10 +3282,7 @@ def test_join_scalar(dtype):
 
 
 def test_join_vector(dtype):
-    values = xr.DataArray(
-        ["a", "bb", "cccc"],
-        dims=["Y"],
-    ).astype(dtype)
+    values = xr.DataArray(["a", "bb", "cccc"], dims=["Y"],).astype(dtype)
 
     targ_blank = xr.DataArray("abbcccc").astype(dtype)
     targ_space = xr.DataArray("a bb cccc").astype(dtype)
@@ -3393,27 +3306,20 @@ def test_join_vector(dtype):
 
 def test_join_2d(dtype):
     values = xr.DataArray(
-        [["a", "bb", "cccc"], ["ddddd", "eeee", "fff"]],
-        dims=["X", "Y"],
+        [["a", "bb", "cccc"], ["ddddd", "eeee", "fff"]], dims=["X", "Y"],
     ).astype(dtype)
 
-    targ_blank_x = xr.DataArray(
-        ["addddd", "bbeeee", "ccccfff"],
-        dims=["Y"],
-    ).astype(dtype)
-    targ_space_x = xr.DataArray(
-        ["a ddddd", "bb eeee", "cccc fff"],
-        dims=["Y"],
-    ).astype(dtype)
+    targ_blank_x = xr.DataArray(["addddd", "bbeeee", "ccccfff"], dims=["Y"],).astype(
+        dtype
+    )
+    targ_space_x = xr.DataArray(["a ddddd", "bb eeee", "cccc fff"], dims=["Y"],).astype(
+        dtype
+    )
 
-    targ_blank_y = xr.DataArray(
-        ["abbcccc", "dddddeeeefff"],
-        dims=["X"],
-    ).astype(dtype)
-    targ_space_y = xr.DataArray(
-        ["a bb cccc", "ddddd eeee fff"],
-        dims=["X"],
-    ).astype(dtype)
+    targ_blank_y = xr.DataArray(["abbcccc", "dddddeeeefff"], dims=["X"],).astype(dtype)
+    targ_space_y = xr.DataArray(["a bb cccc", "ddddd eeee fff"], dims=["X"],).astype(
+        dtype
+    )
 
     res_blank_x = values.str.join(dim="X")
     res_blank_y = values.str.join(dim="Y")
@@ -3438,20 +3344,11 @@ def test_join_2d(dtype):
 
 
 def test_join_broadcast(dtype):
-    values = xr.DataArray(
-        ["a", "bb", "cccc"],
-        dims=["X"],
-    ).astype(dtype)
+    values = xr.DataArray(["a", "bb", "cccc"], dims=["X"],).astype(dtype)
 
-    sep = xr.DataArray(
-        [" ", ", "],
-        dims=["ZZ"],
-    ).astype(dtype)
+    sep = xr.DataArray([" ", ", "], dims=["ZZ"],).astype(dtype)
 
-    expected = xr.DataArray(
-        ["a bb cccc", "a, bb, cccc"],
-        dims=["ZZ"],
-    ).astype(dtype)
+    expected = xr.DataArray(["a bb cccc", "a, bb, cccc"], dims=["ZZ"],).astype(dtype)
 
     res = values.str.join(sep=sep)
 
@@ -3461,8 +3358,7 @@ def test_join_broadcast(dtype):
 
 def test_format_scalar():
     values = xr.DataArray(
-        ["{}.{Y}.{ZZ}", "{},{},{X},{X}", "{X}-{Y}-{ZZ}"],
-        dims=["X"],
+        ["{}.{Y}.{ZZ}", "{},{},{X},{X}", "{X}-{Y}-{ZZ}"], dims=["X"],
     ).astype(np.unicode_)
 
     pos0 = 1
@@ -3474,8 +3370,7 @@ def test_format_scalar():
     W = "NO!"
 
     expected = xr.DataArray(
-        ["1.X.None", "1,1.2,'test','test'", "'test'-X-None"],
-        dims=["X"],
+        ["1.X.None", "1,1.2,'test','test'", "'test'-X-None"], dims=["X"],
     ).astype(np.unicode_)
 
     res = values.str.format(pos0, pos1, pos2, X=X, Y=Y, ZZ=ZZ, W=W)
@@ -3486,17 +3381,13 @@ def test_format_scalar():
 
 def test_format_broadcast():
     values = xr.DataArray(
-        ["{}.{Y}.{ZZ}", "{},{},{X},{X}", "{X}-{Y}-{ZZ}"],
-        dims=["X"],
+        ["{}.{Y}.{ZZ}", "{},{},{X},{X}", "{X}-{Y}-{ZZ}"], dims=["X"],
     ).astype(np.unicode_)
 
     pos0 = 1
     pos1 = 1.2
 
-    pos2 = xr.DataArray(
-        ["2.3", "3.44444"],
-        dims=["YY"],
-    )
+    pos2 = xr.DataArray(["2.3", "3.44444"], dims=["YY"],)
 
     X = "'test'"
     Y = "X"
@@ -3519,18 +3410,16 @@ def test_format_broadcast():
 
 
 def test_mod_scalar():
-    values = xr.DataArray(
-        ["%s.%s.%s", "%s,%s,%s", "%s-%s-%s"],
-        dims=["X"],
-    ).astype(np.unicode_)
+    values = xr.DataArray(["%s.%s.%s", "%s,%s,%s", "%s-%s-%s"], dims=["X"],).astype(
+        np.unicode_
+    )
 
     pos0 = 1
     pos1 = 1.2
     pos2 = "2.3"
 
     expected = xr.DataArray(
-        ["1.1.2.2.3", "1,1.2,2.3", "1-1.2-2.3"],
-        dims=["X"],
+        ["1.1.2.2.3", "1,1.2,2.3", "1-1.2-2.3"], dims=["X"],
     ).astype(np.unicode_)
 
     res = values.str % (pos0, pos1, pos2)
@@ -3541,8 +3430,7 @@ def test_mod_scalar():
 
 def test_mod_dict():
     values = xr.DataArray(
-        ["%(a)s.%(a)s.%(b)s", "%(b)s,%(c)s,%(b)s", "%(c)s-%(b)s-%(a)s"],
-        dims=["X"],
+        ["%(a)s.%(a)s.%(b)s", "%(b)s,%(c)s,%(b)s", "%(c)s-%(b)s-%(a)s"], dims=["X"],
     ).astype(np.unicode_)
 
     a = 1
@@ -3550,8 +3438,7 @@ def test_mod_dict():
     c = "2.3"
 
     expected = xr.DataArray(
-        ["1.1.1.2", "1.2,2.3,1.2", "2.3-1.2-1"],
-        dims=["X"],
+        ["1.1.1.2", "1.2,2.3,1.2", "2.3-1.2-1"], dims=["X"],
     ).astype(np.unicode_)
 
     res = values.str % {"a": a, "b": b, "c": c}
@@ -3561,15 +3448,9 @@ def test_mod_dict():
 
 
 def test_mod_broadcast_single():
-    values = xr.DataArray(
-        ["%s_1", "%s_2", "%s_3"],
-        dims=["X"],
-    ).astype(np.unicode_)
+    values = xr.DataArray(["%s_1", "%s_2", "%s_3"], dims=["X"],).astype(np.unicode_)
 
-    pos = xr.DataArray(
-        ["2.3", "3.44444"],
-        dims=["YY"],
-    )
+    pos = xr.DataArray(["2.3", "3.44444"], dims=["YY"],)
 
     expected = xr.DataArray(
         [["2.3_1", "3.44444_1"], ["2.3_2", "3.44444_2"], ["2.3_3", "3.44444_3"]],
@@ -3583,18 +3464,14 @@ def test_mod_broadcast_single():
 
 
 def test_mod_broadcast_multi():
-    values = xr.DataArray(
-        ["%s.%s.%s", "%s,%s,%s", "%s-%s-%s"],
-        dims=["X"],
-    ).astype(np.unicode_)
+    values = xr.DataArray(["%s.%s.%s", "%s,%s,%s", "%s-%s-%s"], dims=["X"],).astype(
+        np.unicode_
+    )
 
     pos0 = 1
     pos1 = 1.2
 
-    pos2 = xr.DataArray(
-        ["2.3", "3.44444"],
-        dims=["YY"],
-    )
+    pos2 = xr.DataArray(["2.3", "3.44444"], dims=["YY"],)
 
     expected = xr.DataArray(
         [
